Added a verbose flags to the models
Change the initialisation of the means and covariances to use to observed data as a starting point.
In the case of the GMM each mixture's starting values are based on a random subset of the data.

Renamed __impute to __calc_expectation.

Renamed imputed_X to expected_X.

Increased the regularisation on the cov matrix from 1e-6 to 1e-3.

Removed the pre-prcocessing on the input data to normalise it.
This is because for categoricals this might mess things up.

NB: There seems to be some issue with the calculation of LL - the values are lower than one would think from comparing the RMSEs.